Update pom.xml
The provided Maven pom.xml file does not have any syntax errors and should work as intended. However, there are some areas that could be improved or clarified:

Version Specification:

The questdb dependency version is specified as ${project.version}, which refers to the parent project's version. While this works if questdb and this project have the same versioning scheme, it might be clearer to explicitly specify the version.
Dependency Version for questdb:

Check if the questdb version you are using is correct and that 8.0.0-SNAPSHOT exists in your repository. Using SNAPSHOT versions can sometimes lead to issues if the dependency is not properly deployed.
Plugin Configuration:

The maven-surefire-plugin version 2.17 is quite old. Consider updating it to a more recent version to take advantage of bug fixes and new features. The same goes for other plugins like maven-compiler-plugin and maven-jar-plugin.